Fix autocast for meta device

+++ timesead/models/layers/fourier_correlation.py
@@ -257,7 +257,7 @@ class FourierBlock(nn.Module):
         assert Ein == self.Ein
         assert L == self.seq_len
 
-        with torch.autocast(device_type=q.device.type, enabled=False):
+        with torch.autocast(device_type=q.device.type, dtype=torch.float32, enabled=False):
             x = q.float()
             x_ft_c = torch.fft.rfft(x, dim=1, norm=self.fft_norm)  # (B,F,H,Ein)
